backend/python/requirements.txt
flask
supabase
flask_cors
#http file shows in error guid terminal vscode
python-dotenv
